+++ CHANGELOG.md
@@ -1,4 +1,22 @@
+## 0.3.0 (2024-04-29)
 
-## 0.1.0 (2023-09-27)
+Major change in query format. `library` is no longer a provider setting, and the `mssp://` 
+scheme is introduced and required for all queries. Also comes with better tests for the 
+query validation, and improved documentation.
+
+## 0.2.4 (2024-04-29)
+
+First release on PyPI
+
+## 0.2.1 & 0.2.2 (2024-04-29)
+
+Fixes for GitHub actions on releases
+
+## 0.2.0 (2024-04-29)
+
+Decoupled from [snakemake-storage-plugin-http](https://github.com/snakemake/snakemake-storage-plugin-http/) 
+and cleaned up project structure.
+
+## 0.1.0 (2024-04-26)
 
 First version based on the original ([snakemake-storage-plugin-http](https://github.com/snakemake/snakemake-storage-plugin-http/))